test: refactor tests to assure code quality

+++ src/main/java/io/github/gafelix/todo/service/TableRemoverImpl.java
@@ -1,10 +1,13 @@
 package io.github.gafelix.todo.service;
 
 import io.github.gafelix.todo.model.Table;
+import io.github.gafelix.todo.model.User;
 import io.github.gafelix.todo.repository.TableRepository;
 import io.github.gafelix.todo.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.HashSet;
 
 
 @Service
@@ -18,15 +21,20 @@ public class TableRemoverImpl implements TableRemover {
 
     @Override
     public Table deleteTable(String tableId, String userId) {
+        var existentUser = userRepository.findById(userId).orElseThrow();
+        if(!belongsToUser(tableId, existentUser)) throw new IllegalArgumentException();
         var table = tableRepository.findById(tableId).orElseThrow();
-        deleteTableReferenceFromUser(table, userId);
+        deleteTableReferenceFromUser(table, existentUser);
         tableRepository.delete(table);
         return table;
     }
 
-    private void deleteTableReferenceFromUser(Table table, String userId) {
-        var existentUser = userRepository.findById(userId).orElseThrow();
+    private void deleteTableReferenceFromUser(Table table, User existentUser) {
         existentUser.getKnownTablesIds().remove(table.getId());
         userRepository.save(existentUser);
     }
+    private boolean belongsToUser(String tableId, User user) {
+        return (new HashSet<>(user.getKnownTablesIds()).contains(tableId));
+    }
+
 }
